Implement XC scoring algorithm
Far from optimal, and still missing the max score method of Candidate.
Maybe there is a way of not copying objects so much. This should be investigated once tests are running.

+++ include/igclib/pointcollection.hpp
@@ -14,7 +14,7 @@ class PointCollection {
 public:
   PointCollection();
   void insert(const Time &t, const GeoPoint &p);
-  int size() const { return this->points.size(); };
+  size_t size() const { return this->points.size(); };
   void close();
   void box() const;
   timepoints_t::const_iterator begin() const { return this->points.begin(); };
